+++ .travis/mypy
@@ -25,24 +25,15 @@ tmp="$(mktemp -t mypy.XXXX)";
 #   annotations are used, but we ignore the errors for files that are not yet
 #   participating.
 #
-# * These are related to mypy not knowing about zope.interface:
-#   - Incompatible return value type
-#   - Method must have at least one argument
-#   - Unexpected keyword argument (also effects attrs classes' __init__)
-#
 
 mypy "$@"      \
     | grep -v  \
         -e ': error: "Callable\[\[[^]]*\], [^]]*\]" has no attribute "\(skip\|todo\)"'                             \
-        -e ': error: Argument .* has incompatible type "[^"]*"; expected "I[^"]*HTTPHeaders"'                      \
-        -e ': error: Incompatible return value type (got "[^"]*HTTP[^"]*", expected "I[^"]*HTTP[^"]*")'            \
-        -e ': error: Method must have at least one argument'                                                       \
-        -e ': error: Unexpected keyword argument.* for .*"[^"]*HTTP\(Headers\|Request\|Response\)"'                \
         -e ': note: .* defined here'                                                                               \
         -e '^src/klein/_app.py:.[0-9:]*: error: All conditional function variants must have identical signatures'  \
         -e '^src/klein/_app.py:.[0-9:]*: error: Function is missing a type annotation'                             \
         -e '^src/klein/_decorators.py:.[0-9:]*: error: Function is missing a type annotation'                      \
-        -e '^src/klein/_interfaces.py:.[0-9:]*: error: Function is missing a type annotation'                      \
+        -e '^src/klein/_iapp.py:.[0-9:]*: error: Function is missing a type annotation'                            \
         -e '^src/klein/_plating.py:.* error: Function is missing a type annotation'                                \
         -e '^src/klein/_plating.py:.[0-9:]*: error: Need type annotation for '                                     \
         -e '^src/klein/_resource.py:.[0-9:]*: error: Function is missing a type annotation'                        \
